+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Dates are in the DD-MM-YYYY format.
 
+## Unreleased
+
+### Added
+
+- Added a `defaultManager` setting for kickstart, letting users define what manager to use by default when none passed. Defaults to `pnpm` if installed; if not, `npm`.
+
+### Removed
+
+- Removed an unused, undocumented feature (interop layer versioning).
+
 ## [3.2.1] (11-04-2025)
 
 ### Changed